+++ sdk/keyvault/azcertificates/client_test.go
@@ -284,9 +284,10 @@ func TestDisableChallengeResourceVerification(t *testing.T) {
 				},
 				DisableChallengeResourceVerification: test.disableVerify,
 			}
-			client := azcertificates.NewClient(vaultURL, &FakeCredential{}, options)
+			client, err := azcertificates.NewClient(vaultURL, &FakeCredential{}, options)
+			require.NoError(t, err)
 			pager := client.NewListCertificatesPager(nil)
-			_, err := pager.NextPage(context.Background())
+			_, err = pager.NextPage(context.Background())
 			if test.err {
 				require.Error(t, err)
 				require.Contains(t, err.Error(), "challenge resource")