+++ sdk/monitor/azquery/metrics_client_test.go
@@ -45,7 +45,7 @@ func TestQueryResource_BasicQuerySuccess(t *testing.T) {
 func TestNewListMetricDefinitionsPager_Success(t *testing.T) {
 	client := startMetricsTest(t)
 
-	pager := client.NewListMetricDefinitionsPager(resourceURI, nil)
+	pager := client.NewListDefinitionsPager(resourceURI, nil)
 
 	// test if first page is valid
 	if pager.More() {
@@ -66,8 +66,7 @@ func TestNewListMetricDefinitionsPager_Success(t *testing.T) {
 func TestNewListMetricNamespacesPager_Success(t *testing.T) {
 	client := startMetricsTest(t)
 
-	pager := client.NewListMetricNamespacesPager(resourceURI,
-		&azquery.MetricsClientListMetricNamespacesOptions{StartTime: to.Ptr("2022-08-01T15:53:00Z")})
+	pager := client.NewListNamespacesPager(resourceURI, nil)
 
 	// test if first page is valid
 	if pager.More() {